Repository, Service 역할 재부여
